requirements.txt
pyautogui==0.9.54
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
google-generativeai==0.3.2
python-dotenv==1.0.0
pywinauto==0.6.8
pynput==1.7.6
mss==6.1.0
tqdm==4.66.1
pygetwindow==0.0.9
